trying to get rid of the footer on "create-form" page. I will look YT first...
