+++ src/app/(create-form)/layout.tsx
@@ -1,0 +1,5 @@
+const SimpleLayout = ({ children }: { children: React.ReactNode }) => {
+  return <div>Layout</div>
+}
+
+export default SimpleLayout
