--- src/app/create-form/page.scss
@@ -1,1 +1,0 @@
-// CREATE FORM PAGE SCSS
